Remove "declaration of 'lock' shadows a previous local"
On line 182 and 232 the `unlock_guard_t lock(mutex);` shadows the `std::unique_lock<Mutex> lock(*mutex);` on line 152.  In theory it shouldn't be an issue, but I noticed a few `Observed (incorrect) order is: acquisition of lock at` errors in helgrind but they were in asio.  It removed at least 2 of them.  There are a few more but I think those might be helgrind getting confused on a recursive mutex.